requirements.txt
# AI / STT / TTS
openai>=1.0.0
funasr>=1.0.0
modelscope>=1.10.0
torch
torchaudio
SpeechRecognition>=3.10.0
pypinyin>=0.50.0

# Audio Processing
soundfile>=0.12.0
numpy>=1.20.0

# IoT / Networking
paho-mqtt>=1.6.1
